Add changelog document
There is currently no mechanism for kicking off new builds for
cp-translations when community translations are updated.

Adding a changelog should allow translation team to easily create a new
pull request when they know translations need to be added.

This will start the build process and allow us to deploy.